Add BFG edition workaround for options menu crash.
The BFG edition IWADs have their M_GDHIGH lumps changed to say
"Fullscreen:" instead of just "high". This breaks the options menu
which assumes a graphic that is not as wide.

In the same spirit as the title screen workaround, add a workaround
for this as well: use the message "on/off" graphics instead when we
are running using a BFG edition IWAD.

This fixes #341 (thanks Fabian Greffrath).