textscreen: Remove screen size-based font logic.
When the OS is set to a high DPI we want to use the large font. On
Windows we LOGPIXELSX property to do this. Detecting a large screen
size seemed reasonable but probably isn't. Instead, just use the
normal font, but add a TODO to add logic for other operating systems.