pcsound: Extend frequency table to 128 entries.
Extend from 96 to the full 128 entries found in the frequency table in
the Vanilla .exes. This was helpfully posted by Gez to the Doom wiki:

  http://doomwiki.org/wiki/Talk:PC_speaker_sound_effects

Change from the current scheme of storing frequency values to using
the timer divisor values used by the Vanilla .exes; divide into the PC
timer frequency to calculate the frequencies to play.

Thanks to Gez for dumping the full list of frequencies; this fixes #336.