video: Fix crash when running fullscreen with -2.
Remove special logic for setting a scale factor (2x, 3x, etc.) when
running fullscreen instead of windowed. This fixes a crash due to the
fact that I_GraphicsCheckCommandLine() is called before SDL's video
subsystem is initialized.

This makes -2 identical to -geometry 640x480, which is arguably a
better (more easily comprehensible) behavior.

Thanks to Fabian Greffrath for reporting the bug. This fixes #338.