textscreen: Fix jump when clicking on scrollbars.
Clicking on a scrollbar should scroll the cursor to that location, but
the logic for it was buggy because of loss of precision when doing an
integer divide. When dividing by bar_max, seek up to the nearest value
so that the cursor always arrives where we clicked.

Thanks to Alexandre-Xavier for reporting this bug. This fixes #359.