setup: Don't leave joystick subsystem running.
Only init the joystick subsystem when we need to call the joystick
API functions, and quit the subsystem when we are finished. This avoids
conflicts with the joystick widgets that quit the subsystem while the
main code relies on it running.

This fixes a bug where trying to calibrate the joystick twice would
fail on the second attempt.