Use old network sync code by default.
The new netgame client sync code has the potential to be much better
than the original sync code. However, it currently has a buggy
controller and needs revisiting and overhauling (#358).

For the time being, switch back to the old sync code by default and
allow the new code to be enabled using the -newsync parameter
(replacing the -oldsync parameter).