Revert "increase DEFAULT_RAM to 64 MiB"
This reverts commit b24fd9e532119a3a94910c1a6da3dad22ad550d6.

This was a red herring. The game is actually playable with as less as
5 MB of heap memory. However, there is an issue with the memory region
occupied be the DEHACKED patch being Z_Free()'d when the lumpinfo[]
array has been realloc()'d in the mean time. This is simply less
likely to happen when more heap memory is provided in the first place.

See https://github.com/chocolate-doom/chocolate-doom/issues/442