do not fix common mapping errors anymore
After all, they are errors and should be fixed during mapping and not
by the engine. Also, fixing them is a rather drastic deviation from
Vanilla behaviour.

But, more importantly, some mapping errors are demo-critical. However,
due to the order in which variables are set in p_setup.c, it is
impossible to conditionally enable them with one of the "usual
suspects" variables, e.g. demorecording, demoplayback, usergame.
Instead, I had to introduce a new global variable that was once set,
but never reset. So, one ended up with two differently set-up levels
depending on whether the demo loop has been entered from the start
screen or not. Instead of fixing this in each possible location in the
code, I decided to remove the fixes entirely.

# Conflicts:
#	src/crispy.h
#	src/doom/g_game.c
#	src/doom/p_setup.c