make missing textures and missing patches non-fatal
this somewhat non-sensical feature makes it possible to intermix
resources for Doom 1 and Doom 2 in a admittedly very limited scope

at least it does not crash anymore when attempting to load e.g.
maps designed for the 3DO port or load Chex Quest with Doom 2 or
HACX with Doom 1

missing textures are displayed as HOM, i.e. black in Crispy Doom