Video: Set correct texture access mode for the SDL_Texture
SDL_Textures created by SDL_CreateTextureFromSurface will always have
their texture access mode set to SDL_TEXTUREACCESS_STATIC, which means
"changes rarely" and is plain wrong in our case. However, it is
impossible to change this mode once a texture is created.

Now, we create the texture with the correct texture access mode, but
since we are not creating it from the RGBA surface anymore, we have to
set pixel format and dimensions explicitely. We can cope with that. ;-)

TODO: Find out what's the matter with SDL_TEXTUREACCESS_TARGET.