Tweak HACKING style guide.
There was no example for how to write function calls. Do this to
make clear that there should be no space between the function name
and the open paren. Give examples of assignments and if() conditions
as well, and tweak the for() style - there's no reason to omit spaces
around operators for the looping conditions.