Avoid calling strlen() on a potentially unterminated string
The src string may be unterminated and the call to strncpy() be
terminated by reaching dest_size. Instead of calling strlen() on the
src string, check if it has a NUL byte at the same position as the
dest string -- if not, the string was truncated.

Valgrind now gives thumbs up!