net: Include port number in address strings.
When generating string representations of network addresses, include
the UDP port number if it isn't the standard port number. This is
necessary because the string version of the address is used by the
setup tool when filling in the address field; if a non-standard port
is used then it needs to be included.

Also fix byte swapping on the address portion in the same function.

Thanks to Alexandre-Xavier for the bug report on #469.