Get rid of the player2_t type
I introduced this type as a measure for savegame compatibility.

However, I was a bit over-cautious with this step, as the player_t struct
is never saved or loaded in its entirety in saveg_write_player_t() or
saveg_read_player_t(). Thus, it is perfectly safe to add the additional
elements to the player_t struct itself and simply have them ignored
upon game loading and saving.