setup: Add help URLs for configuration windows.
The wiki now has documentation pages for most of the windows in the
setup tool. Add help URLs that link to these.

One other minor change here is that the warp button on the main menu
had to change to F2 instead of F1, which is now the help key.