Make DEH/BEX parser more error-tolerant
If a DEHACKED file does not contain a valid header signature, print a
warning but do not error out.

If a DEHACKED file contains BEX sections which contain non-escaped
newlines and there is no new valid section marker after the newline,
try to proceed parsing with the previous line parser.

Fixes #58