textscreen: Add table cell overflowing.
It's a common pattern in the setup tool that widgets are laid out in
a grid, except there inevitably end up being widgets that do not
conform to this grid. This become ugly to code around, as it ends up
requiring packing tables inside tables to get the desired result, when
all that is really wanted is a widget that can take up more than one
cell in the table. Overflowing allows cells in the table to be
designated as "overflows" of neighboring cells, to facilitate this.