Adjust I_FindScreenMode to be non-static
Make I_FindScreenMode non-static and add a _fullscreen parameter.

I intend to use this routine from v_video as part of the screenshot
code.

Use an underscore prefix to clearly differentiate from the global
variable.

Update all callers. In most cases the last parameter is just passed
the global 'fullscreen', but in a few places pass a static value
instead. This is because I kept doing a double-take when I was
reading 'fullscreen' in the call from a routine that was clearly
only for windowed-modes, and vice-versa.