Swap OPL3 stereo
The OPL3 output to SDL was reversed; All left pans were to the right, and vice versa. 
This is the equivalent of physically reversing the audio patch cables in a mixer.

I have not checked to see if this is deeper in the OPL3 code, but at least on the emulated
OPL3 core this now plays correctly.