remove fullscreen mode selection and related code
With SDL2, we do not change the screen resolution in fullscreen, instead
scaling the game's buffer up to whatever the native resolution is. Therefore,
remove the heuristic for finding a good fullscreen mode, as well as the toggle
between resolutions for fullscreen mode versus windowed.

Based on some prior discussion there are some limited situations where the user
might want chocolate-doom to change the native screen resolution (such as a
resource-constrained machine like a Raspberry Pi running on a moderately high
resolution display like a TV). We should probably provide this as an Advanced
option, but we likely won't need the heuristics for that situation, instead
allowing the user to specify a precise resolution to mode-set.