Search GOG.com installation directories for IWADs
The game is sold on GOG.com now for the Windows platform.  This should
add support for finding the IWADs installed through it.

I've repurposed the Collector's Edition functions for this, the basic
principle for finding the directories is the same between the two, and
it keeps the code duplication low.

resolves #600