Video: Show a disk icon if lumps have been read in the previous tic
This restores the disk icon that is displayed in the bottom right
corner of the screen whenever lumps are loaded. Instead of being drawn
and probably overdrawn again within the same tic, the icon is now
shown for at least one full tic. Also, the icon is not scaled to the
screen in its own step anymore, but is merely projected into
I_VideoBuffer and then scaled with the rest of the scren at the
end of the respective tic.

This closes #603 and lets us get rid up UpdateRect() entirely.