Video: Implement adaptive scaling of the upscaled texture
The texture is always scaled to the next highest integer multiple of
the original resolution whenever the window is resized. If the window
dimensions do not match the original aspect ratio, the rendered area
will be decreased to fit in. Thus, consider the smaller factor for
horizontal and vertical scaling.

Also, allow at least one gametic to pass between two resizes.