Fix rendering of mid-textures with out-of-range screen coordinates
In the MBF code the "spryscale += rw_scalestep;" increment has been moved into the iterator of the for-loop. In the Vanilla code it was at the end of the loop and thus skipped by the "continue;" instruction. This resulted in the whole texture being skipped for rendering as soon as one part of it was out of screen range.

Thanks to vesperas for reporting this rendering glitch.