Fix bug when starting multiplayer Chex Quest games.
The server was rejecting connects when running with chex.wad, because it
was expecting the gamemode to be shareware; in fact, chex.wad contains
four episodes worth of levels (even though only five levels are used).
Technically it is therefore a retail IWAD and we should expect
gamemode=retail when using it. This fixes #647; thanks to Danfun64 for
reporting the bug.