correctly determine the size of the demo lump
Otherwise, in case of longtics demos, one byte of a two-byte angleturn cmd
may be misinterpreted as an early DEMOMARKER byte. Thus, the demosize
variable would be miscalculated and the progress bar exceeds beyond
the framebuffer. Add an extra check to avoid the latter, just in case.

Fixes #85, thanks @plumsinus for reporting!