Fix infinite loop caused by Demon speed bug
Turned out we have been bitten by a classic Vanilla bug. It was
nothing specific to E1M8b and not even to limit-removing or whatsnot.

```
Conversely, it is possible by loading a normal save and starting a
nightmare game repeatedly to continually halve the state durations,
making demons faster and faster. Of particular interest is the chase
state sequence (S_SARG_RUN1 to S_SARG_RUN8), where each of the
individual states that comprise it lasts only two tics. After two
halvings, their duration is reduced to zero, possibly resulting in an
infinite loop as the whole sequence can be looped an infinite amount
of times in a single tic. (This can happen if the demon is awake, but
cannot reach the player; for example, because a door has not finished
opening. Being unable to start attacking, it will continue to chase
forever without any delay, freezing the game as it cannot process
anything else.)
```
http://doomwiki.org/wiki/Demon_speed_bug

Fixes #87

Thanks to @Zodomaniac for the bug report and the savegames!