forbid New Game and (Quick) Load while recording a demo
This means, disable the New Game and Load Game entries from the main
menu and ignore the Quick Load and Load Game function keys in-game.
Play an "Oof" sound when the latter are pressed, a gun shot sound is
already being played when the former are selected, so no other audible
indicator will be heard.