textscreen: Fix handling of mapped input chars
Copy the TranslateKey definition from i_input.c.

This fixes accessing sub-menus such as advanced in the graphics
menu via 'A', for example.

For the future: could we define this (and the supporting stuff)
in just one place? Probably the canonical copy should be in
libtextscreen, since that is included/linked to by chocolate-*,
and not the other way around; but there is already some grey
area e.g. textscreen including doomkeys.h.

This now works as well as the sdl1 setup programs did, I think.
There are still some quirks and things which don't work at all
well if you wander off US-ASCII, just as with the sdl1 version.
I also suspect that the setup programs should respond to changes
in the vanilla_keyboard_mapping setting as part of fixing that.