Refer to the correct sound effect in A_FireCGun
A_FireCGun called S_StartSound with sfx_pistol as an argument.
Therefore the sfx_chgun sound effect is never selected, and the
various code paths that check for whether a sound effect is a
link or not are never executed (sfx_chgun is the only such sound
effect).

I hypothesise that this was changed at some point post Doom 1.2.,
possibly post 1.9 and prior to the linuxdoom source release.

If pitch shifting is turned on, this alters the base pitch for
the chaingun to be shifted upwards. This is consistent with
Doom 1.2's behaviour, although we might need/want to recalibrate
the shifting a little bit further.

Closes: #671.