Set vanilla_keyboard_mapping default to false
I think this is a more appropriate default setting. It means that
user's keymappings will be honoured, which is likely the expected
behaviour on modern systems; it means non-ASCII characters can be
bound to controls, meaning users do not need to know which of the
keys on their keyboard are in the ASCII set and which are not; it
should also play nicer with non-traditional text input devices,
e.g. on-screen keyboards and anything else SDL2 supports (voice-
input?)