add back MUS format header check
Now that Crispy Doom supports so many different file formats for
music, it isn't sufficient anymore to assume that "if it isn't MIDI,
then it must be MUS". Reportedly, the mus2mid() function happily
converted anything that could be assumed to be faintly legit into
MIDI format -- even FLAC files which, of course, would refuse to play
after this conversion. Thus, we add back a check for the "MUS" header
which must be found at the beginning of a music lump in order to
identify it to be in MUS format. This check has been removed because
of the broken D_DM2TTL lump of DECA.WAD which begins with a "PWAD"
header instead of "MUS". Apparently, Vanilla Doom didn't have this
check, but in the other hand only supported MIDI and MUS solely,
anyway.

https://github.com/chocolate-doom/chocolate-doom/commit/784e848b326a6f8b0c1377fc7a1b3c5a945cb009

Sorry, D_DM2TTL lump of DECA.WAD, you won't play anymore, but you are
much too broken with your "PWAD" header for us to detect you as valid
music in MUS format. On the other hand, with this check added back,
all the other music that people throw at Crispy Doom will play
flawlessly.

Thank you @Zodomaniac for the report, fixes #95