Bring back the classic Doom "hit enter" behaviour
When playing PWADs with classic Doom, you always got the "this game
has been modified" banner and had to hit enter to continue.  If you
played a lot of PWADs, you got used to hitting enter regularly.  It
became a core part of the Doom-playing experience. Even if you used
one of those lovely launchers that came on the shovelware CDs which
streamlined the PWAD choosing experience, you always had to hit the
enter key.

This commit restores this classic behaviour and allows those who
still have the muscle-memory "hit enter" behaviour permanently
burned into their nervous systems to finally be at peace.

TODO for a future commit: the coloured top-line of text (red for
1.666, blue for 1.9 iirc)