[sdl2-branch] toggle fullscreen with cmd+enter
This morning's breakfast coding...

Special-case listen for LGUI/RGUI + enter (which is Command on Mac;
Windows key on Windows; Meta or whatever on Linux) and toggle
fullscreen.

On Mac, when you are fullscreen-via-maximize-button, if you push the mouse
pointer to the top edge, the titlebar scrolls down and you could click to
de-maximize. This does not happen if you are fullscreen-via-shortcut. I
think this might be an SDL2 bug or quirk.