remove 256x200 resolution option
This resolution is a scale-down from Doom's native resolution. The vast
majority of chocolate-doom players will not want to use it. Balance out
the resolution table by adding 1920x1440 at the end, instead.

Note it will still be possible for people to use 256x200 if they really
want to: you can supply any width or height in chocolate-doom.cfg.