allow PWADs to provide the lumps necessary for the Episode menu
If the lumps necessary for rendering Doom 2's New Game -> Which
Episode menu are available, the engine looks for the availibility of
the NERVE.WAD PWAD file. Only if this is found, the nervewadfile
variable is set and the menu is actually rendered when selecting New
Game.

Formerly, this was only possible with the BFG Edition's DOOM2.WAD and
was completely skipped when a PWAD was already loaded. The check for
having a PWAD loaded had to be relaxed in order for this to work now,
though. So, things may look weird when another PWAD, e.g. a MegaWAD,
is loaded.