setup: Further simplify display dialog logic.
Previously we tried to find the "nearest" window size when we were
using a window size that was not an integer multiple of the base size.
Instead, create an extra window size radio button when the configured
size is a non-standard one, to preserve the existing configuration.
Also, change all of the code that refers to "screen modes" to now
refer to "window sizes" since that is all we are doing now.