draw bullet puffs in outdoor areas
This implements a P_SpawnPuffSafe() function that spawns a bullet puff
without calling the PRNG by callling P_SpawnMobjSafe() which avoids
this as well. Seems I donÄt need a special version of
P_SetThingPosition() for things which are not registered into the
blockmap anyway.

Fixes #115, thanks ptoing for the idea!