setup: Fixes for SFC30 on Linux systems
The SFC30 joystick presents different names depending on mode (two
bluetooth modes and one USB) and platform (different names for the
same modes on Linux and OS X).  Use glob-matching to catch them all.

Bizzarely, with bluetooth mode on Linux the gamepad presents 16 buttons.
Add another joystick_config_t element so that it is detected properly in
this situation.