make two variables static
The event_t structs in the mouse button and wheel handling routines
have their address passed outside the scope of those routines, which
is undefined behaviour. Mark them as static to address this.

This makes them singletons, meaning no more than one of each type
of event per button/axis can occur in a single tic, but that seems
to be true anyway.

I_ReadMouse and I_HandleKeyboardEvent also have this problem but I
am less confident that applying static is sufficient to fix them.
Mark them with XXX comments to make sure we don't forget.