textscreen: Fix rendering when using a tiling WM.
When using a tiling window manager, the window dimensions as passed
to SDL_CreateWindow() can potentially be ignored, and we can end up
with a different window size even though we didn't ask for the window
to be resizeable.

In this situation, we were scaling the textscreen image to always fill
the window, regardless of size or aspect ratio. This led to some nasty
looking, distorted views in some cases. Instead, always render the
textscreen image at the intended size and center the image within the
window if it is not the requested size.