render weapon sprites centered, but leave actual coordinates untouched
I never really liked the previous implementation which modified the
weapon sprites x-coordinate to render the weapon centered, because
this value is stored in savegames. The new implementation leaves the
actual coordinate value untouched and merely renders the weapon sprite
centered.

This may feel like a regression, because after firing the weapon swing
will continue at its actual position instead of the centered position.
So, while this may feel a bit choppy, the original swing is preserved,
which is much more important to me.