win32: Add cp-with-libs script for Win32 package.
As part of the 3.0 release we want to start bundling many more DLL
libraries than we previously did. We can use the 'objdump' command
to recursively find all DLLs that a Windows binary depends upon. To
ensure that we automatically include any DLLs that are needed and
never miss any, add something analogous to the cp-with-libs script we
already use for building Windows binaries.