video: Always grab mouse when exiting main menu.
When we exit the main menu we should grab the mouse and hide the mouse
cursor, even if the cursor is currently pointing outside of the game
window.

It turned out that the logic to determine whether to grab the mouse was
trying too hard to be well-behaved. While we should release mouse grab
if the app loses focus (so the user can see their mouse cursor again and
click on whatever popup has appeared), we don't need to do the same just
because the mouse cursor is outside the game window's bounds.

This fixes #798. Thanks @CapnClever.