Remove doomfeatures.h and "features" compiles.
This was an idea I had a while back to allow slimmed-down compiles
that would remove certain chocolatey features. As far as I know,
nobody's ever used it, it isn't regularly tested, and it hasn't been
kept up to date as development has progressed. Given this, and
considering that it makes the Makefiles much more complicated to read,
let's just get rid of it.