disable the Cripsness menu in the setup tool
and point players to the in-game Crispness menu instead.

The Crispness menu in the setup tool wasn't able to switch all
features and especially features with more than one possible setting
were just switched on and off, which doesn't do them justice.

Inspired by @JNechaevsky who made it clear how easy it is to miss
certain features in this menu in his suggestion number 2.