video: Add config variable for texture pixel limit.
We use an intermediate texture for scaling up the screen accurately and
limit this to 1600x1200, but for very high resolution displays there can
be some slight blurriness around pixels when they are examined closely.
So provide the ability to override this and use even more memory if
desired.

Thanks to Linguica for the report. This fixes #859.