video: Query SDL API for maximum texture size.
Further tweak to #859. Just to be safe, use the SDL renderer API to
determine the maximum texture dimensions of the graphics hardware,
and don't try to create a texture larger than is supported. As an
additional enhancement, print an informational message when we have
limit the intermediate texture size for any reason.