osx: Tweak colors of command line args box.
In a previous change (I think d8cfdba5bffc8 ?) I changed the text
entry box for command line arguments to be white on a black
background, with the idea being that it was more similar to a DOS
window. @Linguica reports that this doesn't look nice on Macs which
have low-res displays, so revert back to default colors, and also
change the font to use the system font for user-selected fixed
width. I believe this fixes #858.