add more MBF code pointers
That is, all code pointers that MBF supports through its BEX format.
This is the first step on the way to support the SMOOTHED mod.

While at it, introduce a new P_CrossSpecialLinePtr() function that
takes a line_t* pointer as its argument instead of a line number,
required for the A_LineEffect() code pointer.