Second attempt, this time with pipes
Unfortunately, as it turns out, SDL_net does not allow you to create a
TCP listening socket that only listens on localhost.  This has the
unfortunate side-effect of showing a firewall message, which is not
desirable.

Instead, I use pipes to communicate with the subprocess over stdin.
This approach actually works - the current state of the code is if you
launch Chocolate Doom, it will communicate with the subprocess.

This is only the beginning.  Although communication is occurring, I
don't appear to actually be able to hear anything, and Chocolate Doom
crashes as soon as there is an attempt to switch songs.