do not restrict R_FlatNumForName() into the flats range anymore
and while at it, remove the "nwtmerge" workaround

If I remember correctly, this was meant as a workaround to prevent
savegames saved in Crispy Doom with a PWAD loaded from crashing if
restored in Chocolate Doom with the same PWAD loaded with the "-file"
instead of the "-merge" parameter. However, it is an annoying
deviation from upstream and this issue is better dealt with through
documentation. I'll add a sentence to the Wiki:

"If you are going to share savegames between Crispy Doom and Chocolate
Doom, make sure to load all PWADs with the -merge parameter in the
latter."