Support for Watcom C
This adds support for compiling using Watcom C. The main change is the
addition of the PACKED_STRUCT() macro, as Watcom uses a different syntax
to gcc for marking structs as packed. Also, remove a couple of gccisms
that are not supported in Watcom, and the last leftover __WATCOMC__
defines from heretic's d_main.c, which are obsolete anyway.

Thanks to @scfinniss for this work.
# Conflicts:
#	src/doom/doomdata.h