Revert "add a separate key for player respawn"
This reverts commit d8a30da7a84e8bca46141598e8ec73e3102f724c.

As it is now, you may *only* respawn using the dedicated "respawn"
key, not the mouse of joystick. For this to change, I would have to
introduce a dedicated mouse button for "respawn". But this is
impossible, because it is not possible to assign the same mouse button
to different logical actions.

Sorry @Zodomaniac but it seems you have to get used to pressing the
"use" button for respawn. :/