fix revenant internal demo bug
The logic is as follows:

Demo recording is initiated from the command line, so the recording
starts with gametic 0. There is no other possible way to *record* a demo.

When demos are played back from the command line (i.e. by -playdemo or
-timedemo) they do also start with gametic 0 and everything remains in
sync. When demos are played back by the in-game demo loop, however,
they start at arbitrary gametics and we need to substract the gametic
at which the demo playback was started to remain in sync.

@JNechaevsky: This should make it possible to play back entire
speedruns by the in-game demo loop and thus perform a multi-hour
stress test. ;)

# Conflicts:
#	src/doom/g_game.c