Enable SDL_HINT_GRAB_KEYBOARD for special keys
This enables grabbing they keyboard as well as the mouse when
enabling grabs, preventing other software (e.g. the Window Manager)
from filtering key events such as special keys including Print
Screen.

Enable grabs during the key input dialog in setup.

These two changes mean you can bind Print Screen and use it for
actions in-game.

The special keys only work when the grab is enabled: so in-game,
but not in menus: this is a regression to the SDL 1.2 behaviour.

Fixes #765.