remove "screen shot" message from screenshots
This redraws the rendered scene without the obnoxious "screen shot"
message when taking more than one screenshot, because when you are
viewing a screenshot you usually already know that it's a screenshot.
;)

I did this by introducing a new variable "crispy_screenshotmsg". This
variable is set to 2 when a screenshot has been taken. This value
means "we have just taken a screenshot and the 'screen shot' message
will be shown from then next tick on".

In the next tic, this variable is right-shifted in HU_Ticker to become
1 as soon as the "screen shot" message is written into the w_message
widget. This value means "we are currently showing the 'screen shot'
message". When a new message - e.g. a pick-up message - appears, this
variable is right-shifted again in the same function and will then
become 0. It will also get right-shifted to become 0 when the
message_counter runs flat before a new message has appeared. A value
of 0 means that we are currently not showing the 'screen shot'
message.

Now comes the interesting part: If a screenshot is taken while the
variable has value 1 this means we need to get active and remove the
'screen shot' message from the scene before saving the picture. To
achieve this we temporarily raise the variable value to 4 which means
"we are currently taking a screenshot and need to remove the 'screen
shot' message". We do this using the same measures as introduced for
the crispy_cleanscreenshot feature. Once we have finished rendering
the scene we set the variable value back to 2 as usual.

This is probably not the most sophisticated solution, maybe I will
merge this variable and crispy_cleanscreenshot into a {short,short}
struct. But for now it serves the purpose and thus fixes #203, thanks
@Zodomaniac for the idea and discussion.