net: Include protocol list in query responses.
If we try to connect to an incompatible server, the server will reject
the connection attempt. Now that we have a more elaborate scheme to
negotiate a common protocol than just comparing version strings, we
can be more confident about ignoring servers that we definitely won't
be able to connect to. So use the protocol list included in query
responses to exclude query responses from servers we aren't compatible
with.