midiproc: fix music playback of any kind
Unregister songs when they are stopped and wait for an acknowldgement.

Rationale: When SDL_Mixer uses an external library (e.g. fluidsynth)
for music playback, this library may keep the temporary music lump file
open. This means that it cannot be overridden by the next attempt to
register another song. As a consequence, the same song is played over
and over again.

Instead, unregister a song as soon as it is stopped in midiproc and
let the main process wait until the unregistering is acknowledged.
Then it is the roght time to override the music lump file and play
a new song.

Thanks a lot to @Zodomaniac for his willingness and patience to
tackle this issue!