support MUSINFO lump (dynamic music changing)
First working (?) implementation. Kind of limited at the moment,
because only lumps that are known music lumps are currently allowed.
Next step will be to extend support to arbitrary lump names.

This implementation is closely based on PrBoom+'s, but takes care to
support value 0 for the map's default music as well as values up to and
including 64.

Does anyone know a good test map for this feature?

Fixes #128