fix crash when finishing NERVE MAP06 while recording a demo
The fixes a logical error in G_DoWorldDone(). The idea is that
NERVE.WAD behaves like any other ordinary PWAD when recording a demo or
playing a network game (for compatibility with other Choco forks or
ports in general). So, the special casing of gamemission == pack_nerve
does not apply during demos because crispy->singleplayer is false.
Thus, the code jumps into the gamemode == commercial condition which
is still true and tries to show an intermission text after MAP06, just
as it does in Vanilla. However, it cannot find a fitting text string
and crashes. :(

Thanks to @galileo31dos01 for pointing this out on the forums!