video: rename rgbabuffer variable to argbbuffer
This is just a cosmetic change, the variable is static in i_video.c
and its name has been chosen rather arbitrarily based on the false
assumption that ARGB is the most common pixel format nowadays (it is
the standard pixel format for the PNG image format, though).

In Chocolate Doom, we do not hard-code the pixel format anymore, but
rely on the screen's native format instead. Since nowadays all major
systems apparently use the the 8-8-8-8 ARGB format,
"ARGB is the right thing by monopoly vote."
<http://stereopsis.com/doubleblend.html>

So, if we so have to raise expectations about the pixel format by
our choice of the variable name, it should be the most common one.