zero music volume mutes OPL music
Thanks to @JNechaevsky, @Zodomaniac and @mfrancis95 for nagging me
about this and @nukeykt for clarifying OPL chip limitations and
suggesting to pause music when it is muted.

Only caveat: The resume code is always called whenever the music
volume is changed to a non-zero value, but this seems to be harmless
here.

Fixes #170