register up to three buttons at once for lines with more than one switch texture
I know I am a little late to the party, but I think what I am presenting here
is the most Vanilla-friendly possible solution (in terms of code line changes,
that is). It is also savegame compatible with previous code.

Two caveats:
(1) This registers up to three times more buttons for lines
with switch textures than Vanilla code, so in theory this might break
the buttonlist[] limit. This has been mitigated by removing the
MAXBUTTONS limit in two former commits.
(2) This may lead to the same sound getting emitted more than once by the same
line. This is mitigated by the S_StartSoundOnce() function that I
introduced in an earlier commit.

Thanks to @JNechaevsky for spotting this glitch in the Vanilla code
and @bradharding and @jeffdoggett for presenting alternative
approaches -- that I like less than mine, nevertheless. :p