stop "fixing" IWAD maps
Closes #292, thank you very much @Zodomaniac for testing the port
under the most obscure conditions, this really adds to the robustness
of the engine and is highly appreciated!

So, I decided to remove this code. Why?
1) It fiddles around with IWAD maps! While at some point in the past I
might have considered this to be in line with the port's goal of
purism, I generally doubt this decision nowadays.
2) It may be safe for demo recording, but it fails upon demo playback,
because the demoplayback variable is unconditionally reset in g_game.c
before this code is called. It might be possible to fix this, but this
would be a bodge upon a bodge and I want to avoid changing the core
routines in g_game.c just for this.
3) There are PWADs available that fix these two issues and people are
encouraged to use them. It has also shown in the past that the
engine-side fixes were even counter-productive to the intent of these
PWADs.