complete overhaul of the "flipped corpses" feature
This time, instead of maintaining a "white-list" of map things that
may get flipped - even across savegames - add a mobj flag to
explicitly mark sprites as flippable. Then, if an object is flippable
and not shootable (anymore) and its health value is odd (which, of
course, gets randomized) flip its sprite, else don't. This is even
safe across savegames without any further hassle.

This may get extended to further map things, e.g. projectiles or
decoration. Unfortunately, this means one step back in fixing #302,
because a monster's flags and health are reset as soon as it gets
resurrected. Anyway, this fixes #291, thanks @JNechaevsky.