complete overhaul of the Crispness menu page changing
Now we track an index to a CrispnessMenus[] array instead of a pointer
to the current menu. This allows us to simply switch to the next/prev
Crispness menu using PgUp/PgDn (fixes #296) and to wrap around between
the first and last menu page (fixes #305).

Thanks @JNechaevsky and @mfrancis95 for the suggestions!