opl: Unpause when starting a new song.
This appears to match the vanilla (DMX) behavior and some of the higher-
level sound code in s_sound.c appears to depend on it. Specifically,
S_StopMusic() has code that calls I_ResumeSong() if the music is
currently paused. However, S_Start() clears the paused state, which
prevents the call to I_ResumeSong() from occurring. I interpret this as
implying that the call to DMX's I_PlaySong() will resume OPL playback.

Thanks to Fabian and Julia Nechaevsky for noticing the bug.