reset the deathmatch variable when loading a savegame from the menu
Else, when loading a savegame while a deathmatch demo plays in the demo
loop (I am looking at you, AV.WAD), the engine assumes the loaded game is
a deathmatch as well and applies deathmatch rules, i.e. the FRAG widget
in the status bar and respawning from deathmatch spots.

Thanks carburetor and Dwaze for noticing\!