proper sound clipping in Doom 2 MAP08 and The Ultimate Doom E4M8
If there was no idea of Doom 2 when this questionable feature was
implemented, there was definitely no idea of The Ultimate Doom and
its Episode 4 which were released another year later.

Thanks @JNechaevsky for this suggestion. Could you please double-check
my AND/OR logic? This is really getting confusing, thanks! :grin: