replace Choco's README.md with the content of our own homepage
I think we have riped enough to warrant this change. This means that
now we have all the information in one place, i.e. the Github source
repository which is now also the project homepage.

So, I don't have to add a HTML-rendered copy of the old Choco-Wiki
homepage to each release (and a human-readable copy to the Debian
source package). Changes to the contents of this page, especially the
changelog, will now have to go through Pull Requests, just as other
proposed changes to the repo do. @Zodomaniac