Remove long-opt variant of version
Let's stick to single-dash delimited arguments, in common with the
existing ones. Double-dash are a GNU convention for "long options",
typically spelled out (so, -v or --version). We don't use the GNU
conventions in this project. Having the second option is out of
place and potentially causes unnecessary complexity in other places
(shell completion, documentation…)