net: Adjust consistency buffer length for vanilla games.
When sending consistency values we send the player position from
some number of tics in the past. To work with vanilla we need to use
the same circular buffer length as vanilla Doom; however, the new
choco netcode increased the length when BACKUPTICS was changed. For
the same reason we can't change it back without breaking protocol
compatibility. So conditionally change the buffer length if we're
playing a vanilla game.

This now makes vanilla games playable.