net: Fix IPX arbitration ordering.
It is important that when exiting the loop looking for other nodes,
we always send one last packet giving our current state, otherwise
the other nodes may not know that we are ready and will block.

Fixes a bug where we could play against vanilla fine, but two
Chocolate Doom clients connected to a DOSbox IPX server would not
start up properly.