net: Fix address lookups with TCP module.
When resolving an address for TCP communications, it is important
that the resolving net_addr_t is associated with net_tcp_module and
not net_udp_module, otherwise packets will be sent using the wrong
driver. So distinguish between TCP and UDP addresses in the address
table (they're different port namespaces anyway), and factor out
separate functions for NET_UDP_ResolveAddress and
NET_TCP_ResolveAddress.