play attack sound based on state action function (instead of state number)
Fixes #357, thanks @Zodomaniac for the idea

Could you please check (a) if this works as intended with SMOOTHED.wad and
(b) if this feels right with Vanilla Doom? I have the impression that some
sounds come delayed if compared to the original implementation, although
they are spawned in exactly the same state as in-game. Thanks!