setup: Set "Romero Blue" for setup tool.
This uses Romero's "funky blue" color that was used in earlier
versions of the setup tool (and later disabled). I've been using this
in my build of the setup tool for the past few months and kind of like
it; it also gives a fresh look to Chocolate Doom's setup tool.