parametrize colored text escape character
And while at it, let it become the tilde (~). Why? Because it is a
printable ASCII character, so it is possible to print colored texts to
console without accidentally triggering a control sequence. However,
it is non-printable in Doom, because there is no glyph for it in the
HU font. Also, add some safety checks for the colored text rendering
functions.