bring back pickup messages for weapons that are already owned
Weapons that were already owned but picked up again for the ammo did
not trigger a weapon pickup message. This bug got introduced by fixing
the weapon pickup messages for multiplayer games in commit a8da692786.

This reverts the major part of that commit but leaves the intended fix
intact. Fixes #378, thanks @Zodomaniac!