net: Send resend requests to server on deadlock.
The server already triggers resend requests if it detects a potential
deadlock has occurred, but the same scenario can occur if one of the
clients loses enough game data packets from the server. So detect this
on the client side too and trigger artificial resend requests.

Big thanks go to MadDog and Mortrixs for providing network logs of
the deadlock occurring which allowed me to figure out the cause.