fix blinking keys implementation for the classic status bar
This is it, the code is fixed and anyone who may want to use it may do
so. Thanks @Zodomaniac for bringing up this topic again and
@JNechaevsky for confirming my fix works as intended.

I won't be using this code path in Crispy. The classic status bar has
to remain faithful. I am not looking forward to pages of requests in
the Doomworld forums to change this back to Vanilla behaviour.

Fixes #382