net: Increase master server refresh period.
Now that some servers may be behind NAT gateways, it is important that
we send packets to the master server regularly so that it can forward
hole punch requests to us. With the previous value of a refresh every
20 minutes, some NAT gateways may discard the state. Change to 30
seconds (twice per minute) so that the connection is kept alive.

Rationale: Wikipedia article on UDP hole punching says:
> UDP state usually expires after short periods of time in the range
> of tens of seconds to a few minutes

Based on this, 30 seconds seems like a reasonable value.