net: Display appropriate message on connect failure.
Previously, the "reject reason" would not be set if the server simply
never responded to the connection attempt. So ensure that an error
string of some kind is always set when NET_CL_Connect is called.

As part of this, the handling for REJECTED packets has been moved to
the client code; only the server only sends such packets, so it makes
no sense for it to be in the common code.