net: Plumb startup callback to vanilla netcode.
This makes the Hexen startup screen show the ready tics when starting a
vanilla netgame. As part of this, move some code from d_loop.c and refactor
the client code to add a NET_CL_WaitForStart(). This makes the API for the two
network backends a bit more consistent.