INSTALL: Update Steam information and add GOG information.
There was no mention of Strife: Veteran Edition, but it is there now.
There is additional information for installing the games via Steam on
Linux, including Chocolate Doom’s support for detecting IWADs
installed this way.

Doom and Strife are both available on GOG, and can autodetect its
IWADs on Windows, otherwise users will have to install them on Windows
or Wine to get the IWADs first (we don’t have anything in place to
detect Wine installs for either Steam or GOG).