initial support for a fifth episode
Yes, guys, this means support for SIGIL.wad, i.e. the non-compat version.

What is still missing, though, is support for auto-loading it at start-up
and renaming some of the lumps out of the way. Also, I am currently not
sure how to proceed with the stretched SKY5 lump of the Sigil v1.1
release, but this is probably another topic.

Please test the hell out of these changes, with and without the
Buckethead music PWAD loaded.