fixing playing up to three sounds from lines with more than one switch texture
This time, a more conservative approach. It should be possible to
play the switch sound again and again in you hit the same switch
rapidly. Fixes a bug reported by Looper in the forums.

@JNechaevsky Do you still have the test map at hand which we used to
check the initial fix. It should still work as before, but now it
should be possible to "spam" the switch and it should play all sounds.