tolerate up to 1s delay when opening a file for writing
This mitigates a racing condition that may occur when using midiproc
with an advanced audio backend, e.g. fluidsynth with a soundfont for
MIDI rendering.

Before a new song can be played, the main process sends to midiproc
the signal to stop playing the current song. However, it does not
wait for midiproc to succeed with this. Thus, it may happen that the
main process already attempts to override the music lump file on
disc with a new lump, before the previous one has been released by
the music backend. The attempt to open the music file for writing will
fail, but this is actually never checked in the code.

This approach repeatedly tries to open the file for writing for up to
about 1 second, which should be enough time for the music backend to
shut down the previous song and release the file handle.

Reportedly, this fixes #963.

Context: https://github.com/JNechaevsky/russian-doom/issues/127