append WAD name to default savegame name if savegamedir is empty
Apparently, if we have no configuration directory, i.e. on Windows, we
have no savegame directory as well, c.f. M_GetSaveGameDir(). In this
case, all savegames for all IWADs will be saved into the same
directory and there is still risk of mixing them up. Thus, append the
name of the WAD file if savegamedir is empty.

NB: This does only work if the savegame slot is selected by pressing a
joystick button. You are still expected to unambigiously name your
savegames if playing with a keyboard.

Fixes #1180, thanks @Zodomaniac.