bash-completion: always install into $datadir/bash-completion
The bash-completion package by default searches in $HOME/.local and
/usr/local first before /usr, and we should respect local installation
locations instead of always trying to write to a path that is
accessible only to root and could conflict with a package manager.

This removes a toggle in the configure script, which neither actually
worked properly (setting DIR didn’t work) and I believe did the wrong
thing by default anyway.