fix a crash when a door that is actually a plat is opened again while going down
This happens when the "Misc. Sounds Fixes" feature is enabled,
thanks maximanium from the Doomworld forums for the report
and @Zodomaniac for confirmation.

Actually, this is the same bug as fixed in commit 81029ea9,
but has been overlooked so far. So, consider this part 2 of the fix.
Not sure why it only crashed on Windows, though...